Move config to a flux store
