+++ webapp/src/js/components/panoptes/PropertyLegend.js
@@ -4,10 +4,12 @@ import ConfigMixin from 'mixins/ConfigMixin';
 import LegendElement from 'panoptes/LegendElement';
 import {propertyColour, scaleColour} from 'util/Colours';
 import _map from 'lodash/map';
+import FluxMixin from 'mixins/FluxMixin';
 
 let PropertyLegend = React.createClass({
   mixins: [
     PureRenderMixin,
+    FluxMixin,
     ConfigMixin
   ],
