+++ webapp/src/js/components/panoptes/genome/tracks/PerRowIndicatorChannel.js
@@ -44,8 +44,8 @@ let PerRowIndicatorChannel = React.createClass({
         'query'
       ]
     }),
-    ConfigMixin,
     FluxMixin,
+    ConfigMixin,
     DataFetcherMixin('chromosome', 'start', 'end', 'table', 'query', 'width', 'sideWidth', 'colourProperty')
   ],
