+++ webapp/src/js/constants/Constants.js
@@ -34,13 +34,9 @@ let api = keyMirror({
   FETCH_USER: null,
   FETCH_USER_FAIL: null,
   FETCH_USER_SUCCESS: null,
-  STORE_TABLE_QUERY: null,
-  STORE_TABLE_QUERY_FAIL: null,
-  STORE_TABLE_QUERY_SUCCESS: null,
-  DELETE_STORED_TABLE_QUERY_FAIL: null,
-  DELETE_STORED_TABLE_QUERY_SUCCESS: null,
-  SET_DEFAULT_TABLE_QUERY_FAIL: null,
-  SET_DEFAULT_TABLE_QUERY_SUCCESS: null
+  MODIFY_CONFIG: null,
+  MODIFY_CONFIG_SUCCESS: null,
+  MODIFY_CONFIG_FAIL: null
 });
