+++ webapp/src/js/mixins/StoreWatchMixin.js
@@ -4,14 +4,18 @@ let StoreWatchMixin = function() {
   let storeNames = Array.prototype.slice.call(arguments);
   return {
     componentDidMount: function() {
-      let flux = this.props.flux || this.context.flux;
+      if (!this.props.flux && (!this.context || !this.context.flux)) {
+        var namePart = this.constructor.displayName ? " of " + this.constructor.displayName : "";
+        throw new Error('Could not find flux on this.props or this.context' + namePart);
+      }
+      let flux = this.props.flux || (this.context && this.context.flux);
       storeNames.forEach(_bind(function(store) {
         flux.store(store).on('change', this._setStateFromFlux);
       }, this));
     },
 
     componentWillUnmount: function() {
-      let flux = this.props.flux || this.context.flux;
+      let flux = this.props.flux || (this.context && this.context.flux);
       storeNames.forEach(_bind(function(store) {
         flux.store(store).removeListener('change', this._setStateFromFlux);
       }, this));
@@ -30,9 +34,9 @@ let StoreWatchMixin = function() {
 };
 
 StoreWatchMixin.componentWillMount = function() {
-  throw new Error('Fluxxor.StoreWatchMixin is a function that takes one or more ' +
+  throw new Error('StoreWatchMixin is a function that takes one or more ' +
     'store names as parameters and returns the mixin, e.g.: ' +
-    'mixins: [Fluxxor.StoreWatchMixin("Store1", "Store2")]');
+    'mixins: [StoreWatchMixin("Store1", "Store2")]');
 };
 
 module.exports = StoreWatchMixin;
