+++ webapp/src/js/util/Assert.js
@@ -1,8 +1,8 @@
 let Assert = {
-  assertRequired(obj, requiredKey) {
-    requiredKey.forEach((key) => {
-      if (!(key in obj))
-        throw Error(key + ' is a required arg');
+  assertRequired(obj, requiredKeys) {
+    requiredKeys.forEach((key) => {
+      if (!(key in obj) || obj[key] === undefined)
+       throw Error(key + ' is a required arg');
     }
     );
   }
